Revert "ms compiler changes for fft.c"
This reverts commit 99bed07f6476d0243b3981051c087ad5249c5c43.

+++ crypto_kem/hqc-rmrs-128/clean/fft.c
@@ -30,7 +30,7 @@ static void fft_rec(uint16_t *w, uint16_t *f, size_t f_coeffs, uint8_t m, uint32
 static void compute_fft_betas(uint16_t *betas) {
     size_t i;
     for (i = 0; i < PARAM_M - 1; ++i) {
-        betas[i] = (uint16_t) (1 << (PARAM_M - 1 - i));
+        betas[i] = 1 << (PARAM_M - 1 - i);
     }
 }
 
@@ -133,7 +133,7 @@ static void radix_big(uint16_t *f0, uint16_t *f1, const uint16_t *f, uint32_t m_
 
     size_t i, n;
 
-    n = (size_t) (1 << (m_f - 2));
+    n = 1 << (m_f - 2);
     memcpy(Q, f + 3 * n, 2 * n);
     memcpy(Q + n, f + 3 * n, 2 * n);
     memcpy(R, f, 4 * n);
@@ -201,7 +201,7 @@ static void fft_rec(uint16_t *w, uint16_t *f, size_t f_coeffs, uint8_t m, uint32
     // Step 2: compute g
     if (betas[m - 1] != 1) {
         beta_m_pow = 1;
-        x = (size_t) (1 << m_f);
+        x = 1 << m_f;
         for (i = 1; i < x; ++i) {
             beta_m_pow = PQCLEAN_HQCRMRS128_CLEAN_gf_mul(beta_m_pow, betas[m - 1]);
             f[i] = PQCLEAN_HQCRMRS128_CLEAN_gf_mul(beta_m_pow, f[i]);
@@ -223,7 +223,7 @@ static void fft_rec(uint16_t *w, uint16_t *f, size_t f_coeffs, uint8_t m, uint32
     // Step 5
     fft_rec(u, f0, (f_coeffs + 1) / 2, m - 1, m_f - 1, deltas);
 
-    k = (size_t) (1 << ((m - 1) & 0xf)); // &0xf is to let the compiler know that m-1 is small.
+    k = 1 << ((m - 1) & 0xf); // &0xf is to let the compiler know that m-1 is small.
     if (f_coeffs <= 3) { // 3-coefficient polynomial f case: f1 is constant
         w[0] = u[0];
         w[k] = u[0] ^ f1[0];
@@ -299,7 +299,7 @@ void PQCLEAN_HQCRMRS128_CLEAN_fft(uint16_t *w, const uint16_t *f, size_t f_coeff
     fft_rec(u, f0, (f_coeffs + 1) / 2, PARAM_M - 1, PARAM_FFT - 1, deltas);
     fft_rec(v, f1, f_coeffs / 2, PARAM_M - 1, PARAM_FFT - 1, deltas);
 
-    k = (size_t) (1 << (PARAM_M - 1));
+    k = 1 << (PARAM_M - 1);
     // Step 6, 7 and error polynomial computation
     memcpy(w + k, v, 2 * k);
 
@@ -334,7 +334,7 @@ void PQCLEAN_HQCRMRS128_CLEAN_fft_retrieve_error_poly(uint8_t *error, const uint
     compute_fft_betas(gammas);
     compute_subset_sums(gammas_sums, gammas, PARAM_M - 1);
 
-    k = (size_t) (1 << (PARAM_M - 1));
+    k = 1 << (PARAM_M - 1);
     error[0] ^= 1 ^ ((uint16_t) - w[0] >> 15);
     error[0] ^= 1 ^ ((uint16_t) - w[k] >> 15);
